+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @abqm-ds/tokens
 
+## 2.0.11
+
+### Patch Changes
+
+- feat: 🎸 Melhorias em tokens/icons/react
+
+## 2.0.10
+
+### Patch Changes
+
+- feat: 🎸 Melhorias em icons/react/tokens
+
+## 2.0.9
+
+### Patch Changes
+
+- release: 🦋 Nova versão react/icons/tokens
+
 ## 2.0.8
 
 ### Patch Changes